pyproject.toml
[project]
name = "v5_orchestrator"
version = "0.0.1"
description = "Autonomous paper-trading orchestrator for Bybit linear USDT perps."
requires-python = ">=3.12,<3.13"
readme = "README.md"

# Runtime deps kept tight on purpose. Add only what a phase actually needs.
dependencies = [
    "aiohttp>=3.9",
    "httpx[http2]>=0.27",
    "orjson>=3.10",
    "numpy>=1.26",
    "pandas>=2.2",
    "pyarrow>=15",
    "pydantic>=2.6",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-timeout>=2.2",
    "maturin>=1.5",
    "ruff>=0.4",
]
api = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.29",
]
ml = [
    "hmmlearn>=0.3",
    "scikit-learn>=1.4",
]

[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
addopts = "-ra --strict-markers -q"
asyncio_mode = "auto"
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.ruff]
line-length = 100
target-version = "py312"
